backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pandas==2.2.0
openpyxl==3.1.2
edge-tts==6.1.9
pydantic==2.5.3
python-dotenv==1.0.0
aiofiles==23.2.1
boto3==1.34.57
anthropic==0.25.1
